Fixed several bugs (#200)
* fix: Fixed a bug of closing local notebook

* fix: Fixed Chinese link url

* refactor: Better debounce logic

* fix: Fixed deleteNoteDialog memory leak

* feat: Better reference jump in editor mode

* doc: Updated the README.md

+++ src/lib/crossnote.ts
@@ -274,7 +274,9 @@ export default class Crossnote {
 
   public async deleteNotebook(notebookID: string) {
     const notebook = await this.notebookDB.get(notebookID);
-    await pfs.rmdir(notebook.dir);
+    if (!notebook.directoryHandle) {
+      await pfs.rmdir(notebook.dir);
+    }
     await this.notebookDB.remove(notebook);
   }
   public async updateNotebook(notebook: Notebook) {
